1.대시보드에서 미국주식 보여주기 기능추가 2.stock dic에서 sp500 종목 추가

+++ app/src/main/java/com/gomu/gomustock/ui/format/FormatMyStock.java
@@ -7,7 +7,7 @@ public class FormatMyStock {
     public String stock_code;
     public String stock_name;
     public int quantity;
-    public int buy_price;
+    public float buy_price;
     public int cur_price;
     public List<Float> chartdata = new ArrayList<>();
 }
